Rename package mmsmessage to mms
